preempt possible merge conflict
This code is set to be deleted to fix ACE vulns,
so undo the ptrdiff_t change to avoid a merge conflict.

+++ src/cmd/ksh93/sh/init.c
@@ -1290,7 +1290,7 @@ Shell_t *sh_init(int argc,char *argv[], Shinit_f userinit)
 				sh.shpath = sh_strdup(cp);
 			else if(cp = nv_getval(PWDNOD))
 			{
-				ptrdiff_t offset = stktell(sh.stk);
+				int offset = stktell(sh.stk);
 				sfputr(sh.stk,cp,'/');
 				sfputr(sh.stk,argv[0],-1);
 				pathcanon(stkptr(sh.stk,offset),PATH_DOTDOT);
